requirements.txt
fastapi
uvicorn
python-multipart
torch
ultralytics
Pillow
opencv-python-headless